Wrong clock connection for programmer
Serial clock on the ISP6 connector was connected to pin 2 on our ATtiny13. It should have been pin 7.
Pin 2 is for external generation of clock signal. E.g. where you put a crystal. It has nothing to do with
the clock for serial communication which is on pin 7. pin7 is what you use for In System Programming.

Getting this wrong created a very hard to debug error message from avrdude. The same error message you got
for not having any connections to the AVR chip.